chore: Update .env and bdLogica.js files
Update .env file to change the PORT from 3000 to 3067 and add a new variable NOMBRE_EQUIPO with the value "Joomla". In bdLogica.js, update the query for inserting a new user to include the tipo_usuario field and update the query for inserting a new comment to include the atendido field.